+++ src/test/java/com/smarthouse/user/entity/integration/UserRepositoryTest.java
@@ -27,7 +27,7 @@ public class UserRepositoryTest {
     @Test
     public void saveTest(){
         Role user1 = roleRepository.save(new Role(3, "bar", 2));
-        User user = userRepository.save(new User((long) 1,"alex", "uhhu", "009900000", "880000077", user1));
+        User user = userRepository.save(new User( 1,"alex", "uhhu", "009900000", "880000077", user1));
         User fromDb = userRepository.findById(user.getId()).orElse(null);
         assertThat(user.getName()).isEqualTo(fromDb.getName()).isNotNull();
     }
@@ -36,7 +36,7 @@ public class UserRepositoryTest {
     @Test
     public void findByIdTest() {
         Role user1 = roleRepository.save(new Role(3, "bar", 2));
-        User user2 = userRepository.save(new User((long) 1,"alex", "ujju", "009000009", "80000877", user1));
+        User user2 = userRepository.save(new User( 1,"alex", "ujju", "009000009", "80000877", user1));
         User fromDb1 = userRepository.findById(user2.getId()).orElse(null);
         assertThat(user2.getName()).isEqualTo(fromDb1.getName()).isNotNull();
     }
@@ -46,8 +46,8 @@ public class UserRepositoryTest {
     public void deleteTest(){
         Role user1 = roleRepository.save(new Role(1, "bar", 2));
         Role user4 = roleRepository.save(new Role(2, "foo", 2));
-        User user = userRepository.save(new User((long) 6,"alex", "uhhhu", "00900009", "8870007", user1));
-        User user2 = userRepository.save(new User((long) 2,"alex", "hhhhuu", "00000099", "80000877", user4));
+        User user = userRepository.save(new User( 6,"alex", "uhhhu", "00900009", "8870007", user1));
+        User user2 = userRepository.save(new User( 2,"alex", "hhhhuu", "00000099", "80000877", user4));
         userRepository.delete(user);
 
         List<User> allUsers = userRepository.findAll();
@@ -59,8 +59,8 @@ public class UserRepositoryTest {
     public void deleteTestByID() {
         Role user3 = roleRepository.save(new Role(1, "bar", 2));
         Role user4 = roleRepository.save(new Role(2, "foo", 2));
-        User user1 = userRepository.save(new User((long) 3,"alex", "uhhhhu", "009000009", "80000877", user3));
-        User user2 = userRepository.save(new User((long) 2,"alex", "uhhhhu", "00000099", "88000077", user4));
+        User user1 = userRepository.save(new User( 3,"alex", "uhhhhu", "009000009", "80000877", user3));
+        User user2 = userRepository.save(new User( 2,"alex", "uhhhhu", "00000099", "88000077", user4));
         userRepository.deleteById(user1.getId());
 
         List<User> allUsers = userRepository.findAll();
@@ -72,8 +72,8 @@ public class UserRepositoryTest {
     public void findAllTest() {
         Role user3 = roleRepository.save(new Role(1, "bar", 2));
         Role user4 = roleRepository.save(new Role(2, "foo", 2));
-        User alex = userRepository.save(new User((long) 5,"alex", "uuhhhh", "00900009", "88900077", user3));
-        User ron = userRepository.save(new User((long) 6,"alex", "uuhhhh", "00900009", "89000877", user3));
+        User alex = userRepository.save(new User( 5,"alex", "uuhhhh", "00900009", "88900077", user3));
+        User ron = userRepository.save(new User( 6,"alex", "uuhhhh", "00900009", "89000877", user3));
 
         List<User> allUsers = userRepository.findAll();
